Inlined transformation Converted to Long Modified route in order to properly transform and create the final message

+++ src/main/java/org/mycompany/model/InputDataModel.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class InputDataModel {
     String customerId;
     String filename;
-    Integer numberOfHosts;
-    Integer totalDiskSpace;
+    Long numberOfHosts;
+    Long totalDiskSpace;
 }
